// frontend/src/api/models/categorizationRuleRecommendationListResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { CategorizationRuleRecommendationListResponseGeneration } from './categorizationRuleRecommendationListResponseGeneration'
import type { CategorizationRuleSuggestion } from './categorizationRuleSuggestion'

export interface CategorizationRuleRecommendationListResponse {
  generation: CategorizationRuleRecommendationListResponseGeneration
  suggestions: CategorizationRuleSuggestion[]
}